Fix lockup with long-running subscribers
A long-running subscriber would make inbound messages accumulate
in its worker's queue, which would eventually become full and
then block the dispatch thread, deadlocking (assuming the long-
running task was publishing).

The current solution is to avoid sending to full channels and
spread the load, only blocking if *all* workers are full.  This
is fine as long as there are more workers than long-running,
publishing tasks.